file: refactor configc Files APIs
After some experimentation, it was noted that a working program does
not need a lot of the function that was being exposed by the configc's
APIs.  Moreover, the previous API was designed in a way that heavily
relied in heap-allocated data, and assumed that the following sequence
of command was going to be executed:

    API_init()

    API_get()
    if API_diff()
             API_apply()

    API_free()

Furthermore, as the program work its way through the configc's API
flow, information will be loaded and mutate the API structure
accordingly.  While this isn't necessarily bad, it opens up room for
more bug.

Therefore, this commit is the first of series that is refactoring the
current configc's APIs with the intent to make safer, simpler and
more robust.

Experimentation with library consumers, showed a simple software
managing system configuration just requires two functions `diff()` and
`apply()`, in order to implement its basic functionality

The `diff()` is important, as it allow the user to check the state of the
system, and only perform the apply() when required, given that apply(),
in general; will be involve most costly operations.

Also, this major refactor eliminates all heap-allocated data, in favor of
promoting using memory from stack, as well to "force the user" to interact
with the library as data structure, instead of the "more pragmatically"
way.

To exemplify, a code that would have been written like:

    struct file_t *file = file_init("/path/to/file", 06400, 0);

    if (file_get(file) < 0)
            handle_get_error();

    if (file_diff(file))
            if (file_apply(apply) < 0)
                  handle_apply_errors();

    file_free(file);

would become, after this patch; like so:

    struct file_t file = file_t{
    	   .path = "/path/to/file",
	   .mode = 0644,
	   .flags = 0,
    };

    if (file_diff(&file))
	   if (file_apply(&file) < 0)
	           handle_apply_errors();

In the above, the new API promotes the usage of stack allocated
memory, so it would be automatically free()'ed when it goes out of
scope.  Moreover, both file_diff() and file_apply() should also wrap
all the checks and processing it needs without leaking to the users,
as an attempt to make the code more safer.

As summary, here's the following design goals:

  * Attempt to remove any heap-allocated data in favor of more stack-allocated
    friendly API
  * Dynamic data are isolated and not leaked to the API's users
  * Favor fixed-size and bounded operations, again to make the code safer and
    more predictable.

Signed-off-by: Rafael Silva <rafael@raffs.dev>