sysctl: refactor configc sysctl APIs
In a previous commit, Configc's file APIs got refactored in favor of
simpler, safer and more robust code.  As summary, the following goal
was set:

  * Attempt to remove any heap-allocated data in favor of more
    stack-allocated friendly API
  * Dynamic data are isolated and not leaked to the API's users
  * Favor fixed-size and bounded operations, again to make the
    code safer and more predictable.

This commit follows the same foot step and refactor the sysctl APIs.
To exemplify, a code that would have written like so:

    struct sysctl_t *sysctl = sysctl_init(...);

    if sysctl_get(sysctl) == 0
            if sysctl_diff(sysctl) == 0
	            sysctl_apply(sysctl);

    sysctl_free(sysctl)

With the new API, however; the code could look like so:

    struct sysctl_t sysctl = sysctl_t{
            .key = "fs.nr_open",
	    .value = "1234",
    };

    if sysctl_diff(&sysctl)
            sysctl_apply(&sysctl)

The new API promotes stack-allocated memory, encapsulate and avoid
leaking dynamic data structure that is only required when processing
either _diff() or _apply() (e.g. expanding fs.nr_open to "/proc/sys/
fs/nr_open").  Doing so, the intent is to make the code more safer
and simpler for the configc's consumers.

Signed-off-by: Rafael Silva <rafael@raffs.dev>